requirements.txt
transformers>=4.30.0
torch>=1.12.0
accelerate
sentencepiece
pyyaml
streamlit
fastapi
uvicorn
pydantic
